Move DatePicker to a folder
